Major update.
Version is no longer in the package name but in a file
“@version” to ease conflict resolution for git.

All control files are now prefixed with a “@” sign to make
them easily recognisable and to prevent conflict with user
files in future evolutions.

The build is now based on debhelper for better packages.

The “create” command do not put the architecture in the
recipe directory if it is the default (“all”).